requirements.txt
streamlit>=1.25.0
openai>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
